Add bezier and coefficient functions.
Bezier formulas are from
"http://www.13thparallel.org/archive/bezier-curves/". They ultimately
calculate the fastSlowFastStep and slowFastSlowStep functions. They
are not exact as I need exact x values, rather than calculated intermediate
values. But the curve should be sufficient for this purpose. See the graphs
in the excel file.

Coefficient functions calculate a parabola between two points given the
exponent. Changing the exponent changes the steepness of the curves. A
default of 2 is used for the easeIn and easeOut functions.