Feature: Add custom easing function for timelapses movement.
Custom easing allows the user to dampen or steepen (Coefficient curves
only) the curve.

I don't really understand the math of the Bezier curves but placing
either control point at 1.0 and the other at 1.0 - the other control
point results in an S curve either growing Fast -> Slow -> Fast or Slow
-> Fast -> Slow. Now to get the curve to be linear, you decrease the
bigger value to 0.66 and the smaller value to 0.33. In between these
values (1.0 - 0.66) and (0.0 - 0.33) the S curve dampens from fully S
like to linear.